+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+### Changed
+- [INTERNAL] : WIP : prepare the import tool for batch processing of translations.
+
+
 ## [0.2.2] - 2019-09-27
 ### Fixed
 - Initialize correctly vscode-nls